% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{add_yaml}
\alias{add_yaml}
\title{appends a new yaml to the default yaml file, i.e. Sys.getenv("R_DATAYAML")}
\usage{
add_yaml(name = "test", ...)
}
\arguments{
\item{name}{defaults to "test"}

\item{...}{}
}
\value{

}
\description{
appends a new yaml to the default yaml file, i.e. Sys.getenv("R_DATAYAML")
}
