% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_imf.R
\name{BP_sheets}
\alias{BP_sheets}
\title{returns sheets in BP file}
\usage{
BP_sheets(filename = BP2018_download(), search = NULL, and = T)
}
\arguments{
\item{and}{bool - for multiple search strings defaults to TRUE}
}
\description{
returns sheets in BP file
}
