% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{create_drake}
\alias{create_drake}
\title{Title}
\usage{
create_drake(
  name = "000",
  directory = "/Volumes/data/Dropbox/msandifo/documents/programming/r/2018/drake/",
  remove = F
)
}
\arguments{
\item{remove}{}
}
\value{

}
\description{
Title
}
