% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{create_twitter}
\alias{create_twitter}
\title{Title}
\usage{
create_twitter(
  name = "000",
  directory = "/Volumes/data/Dropbox/msandifo/documents/programming/r/twitter/2018/",
  remove = F
)
}
\arguments{
\item{remove}{}
}
\value{

}
\description{
Title
}
