% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemo.R
\name{download_aemo_aggregated}
\alias{download_aemo_aggregated}
\title{Title}
\usage{
download_aemo_aggregated(states = c("NSW", "QLD", "SA", "TAS", "VIC"),
  remote.path = "http://www.nemweb.com.au/mms.GRAPHS/data/",
  local.path = NULL, months = NA, verbose = F, years = NA)
}
\arguments{
\item{years}{}
}
\description{
Title
}
