% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemo.R
\name{download_aemo_current}
\alias{download_aemo_current}
\title{Title}
\usage{
download_aemo_current(states = c("NSW", "QLD", "SA", "TAS", "VIC"),
  remote.path = "http://www.nemweb.com.au/mms.GRAPHS/DATA/",
  local.path = NULL)
}
\arguments{
\item{local.path}{}
}
\description{
Title
}
