% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemo.R
\name{gasbb_ids}
\alias{gasbb_ids}
\title{Title retrieves gasbb  plantypes from AEMO facility file}
\usage{
gasbb_ids(planttype = "PROD",
  gasbb.facility.file = "./data/facility.Rdata")
}
\arguments{
\item{gasbb.facility.file}{location}
}
\description{
Title retrieves gasbb  plantypes from AEMO facility file
}
