% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemo.R
\name{gasbb_zones_delivery}
\alias{gasbb_zones_delivery}
\title{Title}
\usage{
gasbb_zones_delivery(gasbb.prod, zones = c("Roma", "Moomba", "Gippsland",
  "Port Campbell", "Ballera", "Victoria", "Sydney"))
}
\arguments{
\item{zones}{}
}
\description{
Title
}
