% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_agency}
\alias{get_agency}
\title{return  agency for given data set}
\usage{
get_agency(data.set = "cdiac", yaml = Sys.getenv("R_DATAYAML"),
  prepend = "data sourced from")
}
\arguments{
\item{prepend}{}
}
\description{
return  agency for given data set
}
