% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{get_breaks}
\alias{get_breaks}
\title{gets changepoints from sequence  trend}
\usage{
get_breaks(df, Q = 16, minseg = 12, method = "BinSeg")
}
\arguments{
\item{Q}{The maximum number of changepoints to search for using the "BinSeg" method, see changepoint::cpt.mean}

\item{minseg}{minimum lenth of segment}

\item{method}{Choice of "AMOC", "PELT", "SegNeigh" or "BinSeg", see changepoint::cpt.mean}
}
\description{
gets changepoints from sequence  trend
}
