% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_imf.R
\name{get_IMF_BP}
\alias{get_IMF_BP}
\title{Title}
\usage{
get_IMF_BP(my.country = "World", measure = "NGDPRPPPPC",
  years = 1965:2018, sheet = 65, flag.years = NULL, percent = T)
}
\arguments{
\item{percent}{}
}
\description{
Title
}
