% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_info}
\alias{get_info}
\title{returns information for sepcified data set}
\usage{
get_info(data.set = "cdiac", yaml = Sys.getenv("R_DATAYAML"))
}
\arguments{
\item{yaml}{}
}
\value{

}
\description{
returns information for sepcified data set
}
