% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_link}
\alias{get_link}
\title{returns link for given data.set}
\usage{
get_link(data.set = "cdiac", yaml = Sys.getenv("R_DATAYAML"))
}
\arguments{
\item{yaml}{}
}
\value{

}
\description{
returns link for given data.set
}
