% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_name}
\alias{get_name}
\title{returns title for a given data set}
\usage{
get_name(data.set)
}
\arguments{
\item{data.set}{name of dataset}
}
\value{
title
}
\description{
returns title for a given data set
}
