% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_title}
\alias{get_title}
\title{returns title for a given data set}
\usage{
get_title(data.set)
}
\arguments{
\item{data.set}{name of datset}
}
\value{
title
}
\description{
returns title for a given data set
}
