% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_imf.R
\name{IMF2018_download}
\alias{IMF2018_download}
\title{Title}
\usage{
IMF2018_download(local.path = NULL, folder = "IMF", world = FALSE,
  remote.url = "https://www.imf.org/external/pubs/ft/weo/2018/01/weodata/WEOApr2018all.xls")
}
\arguments{
\item{remote.url}{}
}
\description{
Title
}
