% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{list_data_sets}
\alias{list_data_sets}
\title{returns all listed data.sets in yaml file}
\usage{
list_data_sets(yaml = Sys.getenv("R_DATAYAML"), items = NA, select = NULL)
}
\arguments{
\item{select}{}
}
\value{

}
\description{
returns all listed data.sets in yaml file
}
