% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{lng_tm_to_tjd}
\alias{lng_tm_to_tjd}
\title{conversion from lng tonnes per month to TJ per day}
\usage{
lng_tm_to_tjd(n, date)
}
\arguments{
\item{date}{}
}
\description{
conversion from lng tonnes per month to TJ per day
}
