% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ms_yaml_setup}
\alias{ms_yaml_setup}
\title{for msandifo ste up allows copying releavant yaml fiels to another directory with filter}
\usage{
ms_yaml_setup(tmpname = "temp.yaml", outname = NA, filter = c("eia"),
  data.sets = NA)
}
\arguments{
\item{data.sets}{}
}
\description{
for msandifo ste up allows copying releavant yaml fiels to another directory with filter
}
