% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_imf.R
\name{plot_IMF_BP}
\alias{plot_IMF_BP}
\title{Title}
\usage{
plot_IMF_BP(
  df = NA,
  my.country = "World",
  measure = "NGDPRPPPPC",
  years = 1965:2018,
  sheet = 65,
  flag.years = NULL,
  percent = T,
  labels = T,
  repel = T,
  flag.colour = "firebrick3",
  flag.size = 5,
  flag.shape = 18,
  group = NULL
)
}
\arguments{
\item{measure}{}
}
\value{

}
\description{
Title
}
