% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemo.R
\name{read_aemo_aggregated}
\alias{read_aemo_aggregated}
\title{Title}
\usage{
read_aemo_aggregated(local.path = NULL, state = "NSW", files = NULL,
  folder = "aemo")
}
\arguments{
\item{files}{}
}
\description{
Title
}
