% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_data}
\alias{read_data}
\title{reads  data.set into}
\usage{
read_data(
  data.set = "cdiac",
  yaml = Sys.getenv("R_DATAYAML"),
  data = F,
  verbose = F,
  directory = NULL
)
}
\arguments{
\item{verbose}{}
}
\value{
dta.frame usually of form columms
}
\description{
reads  data.set into
}
