% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gladstone_ports.R
\name{read_gladstone_year}
\alias{read_gladstone_year}
\title{Title}
\usage{
read_gladstone_year(years = 2015:lubridate::year(Sys.Date()),
  fuel = "Liquefied Natural Gas", country = "Total")
}
\arguments{
\item{country}{}
}
\description{
Title
}
