% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_imf.R
\name{read_IMF}
\alias{read_IMF}
\title{Title}
\usage{
read_IMF(country = "World", file = IMF2018_download(world = country ==
  "World"), measure = "NGDPRPPPPC", percent = F)
}
\arguments{
\item{percent}{}
}
\description{
Title
}
