% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reorder_dmy}
\alias{reorder_dmy}
\title{reorder_dmy
changes character vector with dates section order dmy to ymd.}
\usage{
reorder_dmy(dates)
}
\arguments{
\item{dates}{character vector assuem to be in ymd version, can include addtional info such as dmy hms}
}
\value{
charcater vector in ymd order
}
\description{
reorder_dmy
changes character vector with dates section order dmy to ymd.
}
