% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{RS_open}
\alias{RS_open}
\title{opens a file in RStudio}
\usage{
RS_open(fn = Sys.getenv("R_DATAYAML"))
}
\arguments{
\item{fn}{defults to Sys.getenv("R_DATAYAML")}
}
\value{

}
\description{
opens a file in RStudio
}
\examples{
RS_open(Sys.getenv("R_DATAYAML"))
}
