% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{set_month_day}
\alias{set_month_day}
\title{sets  day of month in date to day}
\usage{
set_month_day(date, day = 15)
}
\arguments{
\item{date}{date or datetime}

\item{day}{numeric}
}
\value{

}
\description{
sets  day of month in date to day
}
\examples{
reproscir::set_month_day(lubridate::ymd("2012-01-01"),15)
#"2012-01-15"
}
