% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_yaml}
\alias{set_yaml}
\title{Title}
\usage{
set_yaml(
  update,
  data.set = "eia.drygas",
  field = "link",
  yaml = Sys.getenv("R_DATAYAML")
)
}
\arguments{
\item{yaml}{}
}
\value{

}
\description{
Title
}
