% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_yaml_link}
\alias{set_yaml_link}
\title{set_yaml_link}
\usage{
set_yaml_link(update, data.set = "eia.drygas", yaml = Sys.getenv("R_DATAYAML"))
}
\arguments{
\item{update}{new link}

\item{yaml}{file}
}
\value{
New link as inserted in to yaml file
}
\description{
set_yaml_link
}
\examples{
set_yaml_link("https://www.eia.gov/naturalgas/weekly/img/shale_gas_201807.xlsx", data.set="eia.drygas")
}
