% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Sys.open}
\alias{Sys.open}
\title{Title}
\usage{
Sys.open(fn = "~/.Renviron", editor = Sys.getenv("R_EDITOR"))
}
\arguments{
\item{fn}{filname}

\item{editor}{as set by "R_EDITOR envirnment variable as set in .R}
}
\value{
side effect is open a
}
\description{
Title
}
\examples{


}
