% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{tjday_to_mw}
\alias{tjday_to_mw}
\title{conversion of terajoules per day to megawatts}
\usage{
tjday_to_mw(input)
}
\arguments{
\item{n}{terajoules per day
a common "unit" of  gas production in Australia}
}
\value{
numeric
}
\description{
conversion of terajoules per day to megawatts
}
\examples{
tjday_to_mw(30)
}
