% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{update_all_data}
\alias{update_all_data}
\title{updtses all data sets in yaml file}
\usage{
update_all_data(yaml = Sys.getenv("R_DATAYAML"), items = NA)
}
\arguments{
\item{items}{}
}
\description{
updtses all data sets in yaml file
}
