% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{update_data}
\alias{update_data}
\title{updates data ste by downloading latest version}
\usage{
update_data(data.set = "cdiac", yaml = Sys.getenv("R_DATAYAML"),
  verbose = T)
}
\arguments{
\item{verbose}{}
}
\description{
updates data ste by downloading latest version
}
