% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gladstone_ports.R
\name{update_gladstone}
\alias{update_gladstone}
\title{Title}
\usage{
update_gladstone(local.path = NULL,
  years = 2015:lubridate::year(Sys.Date()), update = T)
}
\arguments{
\item{years}{}
}
\description{
Title
}
