% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validate_directory}
\alias{validate_directory}
\title{copied directly from smapr}
\usage{
validate_directory(destination_directory, folder = "aemo")
}
\arguments{
\item{folder}{}
}
\value{

}
\description{
copied directly from smapr
}
