% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_imf.R
\name{BP2018_download}
\alias{BP2018_download}
\title{Title}
\usage{
BP2018_download(
  local.path = NULL,
  folder = "BP",
 
    remote.url = "https://www.bp.com/content/dam/bp/business-sites/en/global/corporate/xlsx/energy-economics/statistical-review/bp-stats-review-2018-all-data.xlsx"
)
}
\arguments{
\item{remote.url}{}
}
\value{

}
\description{
Title
}
