% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_imf.R
\name{BP_all}
\alias{BP_all}
\title{Title}
\usage{
BP_all(filename = BP2018_download(), sheet = 32, countries = NA,
  years = 1965:2017, na.rm = TRUE, verbose = F, fuel = NA,
  units = NA)
}
\arguments{
\item{units}{}
}
\description{
Title
}
\examples{
(BP_sheets( search=c("Oil.*Prod.*To"))$sheet \%>\%
BP_all( sheet=., count=c("Aus", "Ven", "Russ", "US", "Sau")))$data \%>\%
ggplot(aes(year, value, col=region))+
geom_line()
}
