% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_imf.R
\name{plot_IMF_BP}
\alias{plot_IMF_BP}
\title{Title}
\usage{
plot_IMF_BP(my.country = "World", measure = "NGDPRPPPPC",
  years = 1965:2017, sheet = 57, flag.years = NULL, percent = T)
}
\arguments{
\item{measure}{}
}
\description{
Title
}
